Created by svn2p4.pl. ------------------------------------------------------------------------ r73 | awk | 2007-06-07 02:59:31 -0400 (Thu, 07 Jun 2007) | 3 lines Changed paths:    M /recsched/trunk/MainWindowController.m    M /recsched/trunk/XTVDParser.h    M /recsched/trunk/XTVDParser.m    M /recsched/trunk/recsched.xcodeproj/project.pbxproj    M /recsched/trunk/recsched_AppDelegate.m
Parsing uses an instance of a thread object not the class.
Respond to the NSManagedObjectContextSavedNotification by sending a set of both the new and updated objects to the app delegate. The app delegate will use the object IDs from the set to fault the objects into the main threads object context, and then post a second RSManagedObjectContextUpdated notification.
MainWindowController responds to the RSManagedObjectContextUpdated notification by sending prepareContent to the lineups array controller of the main window which triggers the array to be updated with the new items (and as a result update the UI).
------------------------------------------------------------------------

[git-p4: depot-paths = "//recsched/recsched/main/": change = 102]