Created by svn2p4.pl. ------------------------------------------------------------------------ r194 | awk | 2007-09-14 16:30:53 -0400 (Fri, 14 Sep 2007) | 6 lines Changed paths:    M /recsched/trunk/English.lproj/MainMenu.nib/designable.nib    M /recsched/trunk/English.lproj/MainMenu.nib/keyedobjects.nib    M /recsched/trunk/MainWindowController.m    M /recsched/trunk/Preferences.m    M /recsched/trunk/RSCommonAppDelegate.h    M /recsched/trunk/RSCommonAppDelegate.m    M /recsched/trunk/RecSchedProtocol.h    M /recsched/trunk/RecSchedServer.m    M /recsched/trunk/recsched_AppDelegate.h    M /recsched/trunk/recsched_AppDelegate.m    M /recsched/trunk/recsched_bkgd_AppDelegate.h    M /recsched/trunk/recsched_bkgd_AppDelegate.m
Introduce a USE_SYNCSERVICES #define in RSCommonAppDelegate.h (currently off).
If USE_SYNCSERVICES is NOT defined then the app will open the background servers persistent store in read only mode and will
do no saving or syncing of it's own store. The background app will also not perform any syncing.
I think this may be more reliable and certainly faster than using SyncServices to keep the two stores in sync - mechanisms do need to be implemented however to handle the 'pushing' of tuner/channel details to the background app and also to make sure that the foreground app is notified to update when appropriate. Initial testing seems to indicate that just causing a refresh (change current time in the schedule grid) is enough to refetch new data from the readonly background store after it's been updated with new schedule data.


------------------------------------------------------------------------

[git-p4: depot-paths = "//recsched/recsched/main/": change = 224]