Created by svn2p4.pl. ------------------------------------------------------------------------ r213 | awk | 2007-10-04 00:39:38 -0400 (Thu, 04 Oct 2007) | 6 lines Changed paths:    M /recsched/trunk/RecSchedServer.m    M /recsched/trunk/RecordingThread.h    M /recsched/trunk/RecordingThread.m    M /recsched/trunk/recsched_bkgd_AppDelegate.m
Initialze the recording thread with the RecSchedServer object.
When recording starts use the uiActivity object on the server to create an update an activity with progress information.
Use NSData and NSFileManager, NSFileHandle to recieve and write the video data to a file.
Locate the file in the Users Movies folder and name it with the program ID, Title and Subtitle


------------------------------------------------------------------------

[git-p4: depot-paths = "//recsched/recsched/main/": change = 243]